Fix S2184 and S2104 (#2328)
* Fix SA1204

* Fix warning S2184

* Remove S2184 and SA1204 warning supressions

+++ test/Polly.Specs/Retry/WaitAndRetryAsyncSpecs.cs
@@ -1070,7 +1070,7 @@ public class WaitAndRetryAsyncSpecs : IDisposable
         attemptsInvoked.Should().Be(1);
 
         watch.Elapsed.Should().BeLessThan(retryDelay);
-        watch.Elapsed.Should().BeCloseTo(shimTimeSpan, precision: TimeSpan.FromMilliseconds((int)shimTimeSpan.TotalMilliseconds / 2));  // Consider increasing shimTimeSpan, or loosening precision, if test fails transiently in different environments.
+        watch.Elapsed.Should().BeCloseTo(shimTimeSpan, precision: TimeSpan.FromMilliseconds(shimTimeSpan.TotalMilliseconds / 2));  // Consider increasing shimTimeSpan, or loosening precision, if test fails transiently in different environments.
     }
 
     [Fact]
